+++ src/test/java/kitchenpos/order/application/InMemoryOrderRepository.java
@@ -1,8 +1,8 @@
-package kitchenpos.application;
+package kitchenpos.order.application;
 
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 
 import java.util.*;
