+++ src/test/java/kitchenpos/ordertable/application/OrderTableServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.ordertable.application;
 
-import kitchenpos.application.InMemoryOrderRepository;
-import kitchenpos.domain.OrderRepository;
-import kitchenpos.domain.OrderStatus;
+import kitchenpos.order.application.InMemoryOrderRepository;
+import kitchenpos.order.domain.OrderRepository;
+import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.ordertable.domain.OrderTable;
 import kitchenpos.ordertable.domain.OrderTableRepository;
 import org.junit.jupiter.api.BeforeEach;
